requirements.txt
# ===========================================
# CORE LIBRARIES
# ===========================================
numpy>=1.26
pandas>=2.2
pyarrow>=17.0.0        # parquet, arrow, fast I/O
polars>=1.9.0          # alternativa más rápida a pandas para parquet
duckdb>=1.0.0          # queries SQL sobre parquet

# ===========================================
# DATA INGESTION & EXCHANGE
# ===========================================
requests>=2.32.3
python-dateutil>=2.9.0
ccxt>=4.3.89            # conexión unificada exchanges (Spot/Futuros)
websocket-client>=1.8.0 # live WS feeds

# ===========================================
# ANALYSIS & TECHNICAL INDICATORS
# ===========================================
ta-lib-bin>=0.4.28; platform_system != "Windows"
TA-Lib>=0.4.28; platform_system == "Windows"
# (En Windows puede requerir la wheel precompilada de TA-Lib)
# alternativa si falla: finta>=1.3, o implementar indicadores a mano

# ===========================================
# MACHINE LEARNING / RL
# ===========================================
gymnasium>=0.29.1
stable-baselines3>=2.3.2
torch>=2.3.0
tensorboard>=2.17.0
scikit-learn>=1.5.1
optuna>=3.6.1           # optimización de hiperparámetros

# ===========================================
# CONFIG & UTILITIES
# ===========================================
PyYAML>=6.0.2
pydantic>=2.9.2
rich>=13.8.0
tqdm>=4.66.5

# ===========================================
# TESTING & DEV
# ===========================================
pytest>=8.3.2
black>=24.8.0
ruff>=0.6.9
mypy>=1.11.2
